* Fix bug where many elements in the code completion popup did not show up as private (even though they are showing up as private in the navigation outline)
* Fix a bug where trying to ctrl-click for Go To Declaration into an item that comes from the
preindexed documentation (Yahoo, JQuery, Dojo) would give a misleading error message
(it would say "Foo is a built-in method and has no JavaScript source"). This fix displays
the more accurate "Cannot find the source for Foo". It also fixes an NPE that might occur
if the documentation for a function is not found.

* Fix a possible NPE which can happen when you're doing code completion on some index entries. The problem
was that you could end up with searching for both name=null,in=null which is invalid.

* Fix bug where code completion was showing keywords in many places they are not valid. For example,
doing completion like this:
   x = new |
or
   y.|
should not show keywords like while, for, do, etc. - they are clearly not allowed in these places.

Also, fix new completion (the first case above) when the caret is at the end of the line (right in
front of a \n/EOL).

* Fix bug where "new" completion (which lists possible constructors) would not include many classes
that have default constructors. These aren't stored as contructor methods in the index; they are
stored as properties, so update the index search to handle this.

* Fix a couple of bugs in indexing.
First, @attribute is synonymous with @property, and occurs in some of the YUI documentation.
This is covered by the updated unit test JsIndexerTest.testScriptDoc2

Also, fix a bug in extracting data from the sdoc.zip file; it was computing incorrect filename URLs
when recursing through the file structure. It should be using FileObject.getNameExt(), not FileObject.getName(),
when creating a URL; without this, the stored fileurls were missing their .sdoc extension and documents
couldn't be found.

* Fix bug in error parser recovery.
I based this for JavaScript from the similar error recovery in Ruby -- but I forgot to update the
comment delimiter from "#" to "//"...

* 131667: Overly aggressive "code has no side effects" warning
Fix the above bug: suppress warnings on Token.EMPTY nodes in HTML files. These are added into
the HTML file as part of the HTML to JavaScript embedding model creation (I add some extra semicolons
to prevent accidentally continuing statements from one <script> block to the next.).

This fix also corrects a bug in offset computation where the trailing comma quickfix would
show up on the wrong line in HTML files - because I had forgotten to do an AST to lexer based position
translation which always has to be done or things don't work in embedded files.

* Fix bug in the unsupported calls quickfix. It wasn't able to compute the call type for AST expressions
where you had nested function calls. This was because I was basing the hint on the Token.NAME node instead
of the correct Token.STRING node which is always present for calls; the Token.NAME node is only there
for fully-qualified function call names. I also changed the quickfix popup dialog title from listing
the bit long "Function {0} is not supported by all of your targeted browsers" which caused wide dialogs,
to just the fully qualified name of the function.

* The Trailing Comma warning should have severity error, not warning - since it's an error you
really need to deal with.

* Modify the Rhino AST for switch-statements. Rhino would rewrite the AST quite a bit where it no
longer reflects the source structure. This is bad for the IDE in a couple of ways. Martin
needs this fixed for the AST-based formatter.

* 131665: NPE in GSF
This is just a documentation fix to gsf's javadoc to clarify the usage of an API which was
called with invalid params from PHP.

* 131671: NPE in GSF
This is a fix for a scenario which has happened to some users where GSF tries to index a file
which is no longer there -- and the file is named something like ".#foo". These aren't files
we shouldn't try to index anyway - they are shortlived files typically created a backup files
by various editors and/or version control systems.


(Also update lots of unit tests for the above fixes)