141001: Try and isolate the DataLoader API usage in GSF
Apply patch from Emilian Bold which adds a level of indirection to a lot
of DataObject access in GSF. This makes it easier to support GSF
languages in some other contexts.  It hopefully will also make it easier
in the future to support file-less editing (e.g. JavaScript in a property sheet field etc.)