Fix a bug where the hint to rename a local variable (with wrong camel case conventions, e.g. argList=>arg_list) didn't include the surrounding method's parameter list. I found the same bug in a couple of other local name related quickfixes (nested block vars, block var reuse) and fixed it there too.
Then I noticed that these fixes (interactive quickfixes which invoke instant-renaming mode) didn't have unit tests of the fixes themselves, only the part which finds the problem. Now that I have preview capability, I modified the GSF infrastructure to also support applying quickfixes for hints that implement the PreviewableFix interface; in that case, it will use the generated preview to generate the golden file. Clean up the introduce method/variable tests to use this mechanism instead of the old way which would use a variable to side effect set the target name.