Don't try to add non Python interpreter as Python platform
In PythonPlatformManager, if the commands succeed but is not a Python interpreter
its name will be null (eg /usr/bin/ls). This will cause subsequent call to
"Manage" to fail due to a NullPtrException. We prevent this by returning a null
platform if the name is null.

This can happen if the user enters a wrong path when registering a new Python
platform.